dhcpv6: Fix issue with time_t and format string
  CC       src/connmand-dhcpv6.o
In file included from src/connman.h:139,
                 from src/dhcpv6.c:39:
src/dhcpv6.c: In function ‘__connman_dhcpv6_start_renew’:
./include/connman/log.h:146:39: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 7 has type ‘time_t’ {aka ‘long long int’} [-Wformat=]
  146 |                         connman_debug("%s:%s() " fmt, \
      |                                       ^~~~~~~~~~
src/dhcpv6.c:1434:9: note: in expansion of macro ‘DBG’
 1434 |         DBG("T1 %u T2 %u expires %lu current %lu started %lu", T1, T2,
      |         ^~~
./include/connman/log.h:146:39: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 8 has type ‘time_t’ {aka ‘long long int’} [-Wformat=]
  146 |                         connman_debug("%s:%s() " fmt, \
      |                                       ^~~~~~~~~~
src/dhcpv6.c:1434:9: note: in expansion of macro ‘DBG’
 1434 |         DBG("T1 %u T2 %u expires %lu current %lu started %lu", T1, T2,
      |         ^~~
src/dhcpv6.c: In function ‘__connman_dhcpv6_start_pd_renew’:
./include/connman/log.h:146:39: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 6 has type ‘time_t’ {aka ‘long long int’} [-Wformat=]
  146 |                         connman_debug("%s:%s() " fmt, \
      |                                       ^~~~~~~~~~
src/dhcpv6.c:2318:9: note: in expansion of macro ‘DBG’
 2318 |         DBG("T1 %u T2 %u expires %lu current %lu started %lu", T1, T2,
      |         ^~~
./include/connman/log.h:146:39: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 7 has type ‘time_t’ {aka ‘long long int’} [-Wformat=]
  146 |                         connman_debug("%s:%s() " fmt, \
      |                                       ^~~~~~~~~~
src/dhcpv6.c:2318:9: note: in expansion of macro ‘DBG’
 2318 |         DBG("T1 %u T2 %u expires %lu current %lu started %lu", T1, T2,
      |         ^~~
./include/connman/log.h:146:39: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 8 has type ‘time_t’ {aka ‘long long int’} [-Wformat=]
  146 |                         connman_debug("%s:%s() " fmt, \
      |                                       ^~~~~~~~~~
src/dhcpv6.c:2318:9: note: in expansion of macro ‘DBG’
 2318 |         DBG("T1 %u T2 %u expires %lu current %lu started %lu", T1, T2,

+++ src/dhcpv6.c
@@ -1431,8 +1431,8 @@ int __connman_dhcpv6_start_renew(struct connman_network *network,
 
 	current = time(NULL);
 
-	DBG("T1 %u T2 %u expires %lu current %lu started %lu", T1, T2,
-		(unsigned long)expired, current, started);
+	DBG("T1 %u T2 %u expires %"PRId64" current %"PRId64" started %"PRId64"",
+					T1, T2, expired, current, started);
 
 	if (T1 == 0xffffffff)
 		/* RFC 3315, 22.4 */
@@ -2315,8 +2315,8 @@ int __connman_dhcpv6_start_pd_renew(struct connman_network *network,
 
 	current = time(NULL);
 
-	DBG("T1 %u T2 %u expires %lu current %lu started %lu", T1, T2,
-		expired, current, started);
+	DBG("T1 %u T2 %u expires %"PRId64" current %"PRId64" started %"PRId64"",
+					T1, T2, expired, current, started);
 
 	if (T1 == 0xffffffff)
 		/* RFC 3633, ch 9 */
